requirements.txt
numpy
pandas
streamlit
faiss-cpu
ragatouille